Moving source from Kenai to Github.
Changed IDE from Eclipse to ItelliJ because I like it.

Updated .gitignore to github suggestions.

Re-added DailyScheduleAdapter. This class was never added to source control
on the last few commits.

Planning

The schedule persistence will be changed from serialization to a database.
This will hopefully fix the issue with schedule loss.

Rebranding will occur. This will include a new icon and color scheme which is
already in place.

Possibly considering adding a Marketing framework.